tests/wakelock/WakeLock-object-is-independent.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WakeLock independent Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

promise_test(async t => {
  let iframe = document.createElement('iframe');
  iframe.srcdoc = '<script>' +
                  '  let iframeWakeLock, iframeRequest;' +
                  '  window.onmessage = async message => {' +
                  '    if (message.data === "INIT") {' +
                  '      iframeWakeLock = await navigator.getWakeLock("screen");' +
                  '      parent.postMessage(iframeWakeLock.active, "*");' +
                  '    } else if (message.data === "ACQUIRED") {' +
                  '      iframeRequest = iframeWakeLock.createRequest();' +
                  '      parent.postMessage(iframeWakeLock.active, "*");' +
                  '    } else if (message.data === "RELEASED") {' +
                  '      iframeRequest.cancel();' +
                  '      parent.postMessage(iframeWakeLock.active, "*");' +
                  '    }' +
                  '  };' +
                  '<\/script>';
  
  let wakeLock1 = await navigator.getWakeLock("screen");
  let request1 = wakeLock1.createRequest();
  let wakeLock2 = await navigator.getWakeLock("screen");
  iframe.onload = () => iframe.contentWindow.postMessage('INIT', '*');
  document.body.appendChild(iframe);
  let initMessage = await window.onmessage;
  assert_true(wakeLock1.active, "the active is true if the wake lock has been acquired");
  assert_false(wakeLock2.active, "the active is false before the wake lock is acquired");
  assert_false(initMessage, "the active is false before the iframe wake lock is acquired");

  iframe.contentWindow.postMessage('ACQUIRED', '*');
  let acquiredMessage = await window.onmessage;
  assert_true(wakeLock1.active, "the active is true if the wake lock has been acquired");
  assert_false(wakeLock2.active, "the active is false before the wake lock is acquired");
  assert_true(acquiredMessage, "the active is true if the iframe wake lock has been acquired");

  request1.cancel();
  let request2 = wakeLock2.createRequest();
  iframe.contentWindow.postMessage('RELEASED', '*');
  let releasedMessage = await window.onmessage;
  assert_false(wakeLock1.active, "the active is false if the wake lock has been released");
  assert_true(wakeLock2.active, "the active is true if the wake lock has been acquired");
  assert_false(acquiredMessage, "the active is false if the iframe wake lock has been released");
}, "Test that the WakeLock active is independent.");

</script>
